+++ test/main/stream_server.cpp
@@ -125,6 +125,12 @@ esp_err_t StreamServerStream::SetWriteTimeout(TickType_t timeout) {
 
 //==============================================================================
 
+StreamServer::~StreamServer() {
+  StopTask();
+}
+
+//==============================================================================
+
 esp_err_t StreamServer::HandleRequest(PL::Stream& stream) {
   ESP_RETURN_ON_ERROR(stream.Read(receivedData, sizeof(receivedData)), TAG, "stream read failed");
   if (receivedData[0] == disableDataToSend[0])