+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - StreamServer task not being stopped before the derived object is destroyed.
+- LockGuard multiple object constructor parameter documentation.
 
 ## [2.0.1] - 2026-08-19
 ### Fixed